move HierarchicalSwarmDashboard out of hiearchical_swarm.py into its existing utils module
